+++ src/userScript.js
@@ -2,7 +2,7 @@ import 'whatwg-fetch';
 import './domrect-polyfill';
 
 import { handleLaunch } from './utils';
-import { WebOSVersion, initializeWebOSVersion } from './webos-utils.js';
+import { WebOSVersion } from './webos-utils.js';
 
 import { initBlockWebOSCast } from './block-webos-cast'; 
 
@@ -18,7 +18,7 @@ document.addEventListener(
 import './adblock.js';
 import './sponsorblock.js';
 import './font-fix.css';
-import './thumbnail-quality';
+import './thumbnail-quality.js';
 import './screensaver-fix';
 import './yt-fixes.css';
 import './watch.js';